Translations moved to own comp. Meta tag added.
