+++ package.json
@@ -20,6 +20,7 @@
 		"react-dom": "^16.13.1"
 	},
 	"devDependencies": {
+		"@tailwindcss/custom-forms": "^0.2.1",
 		"json-loader": "^0.5.7",
 		"postcss-flexbugs-fixes": "4.2.1",
 		"postcss-preset-env": "^6.7.0",
